新增收款單  後台資料   修改銷貨消退方法

+++ grails-app/domain/foodpaint/AccountSheet.groovy
@@ -52,7 +52,7 @@ class AccountSheet{
      /**
      * 匯率
      */
-    Double rate
+    long rate
     
      /**
      * 幣別
